Change todoStatus DOM to checkbox and fix priority
todoStatus in dom.js is now a checkbox so that the status of a todo can
be changed.

Priority used to have the value of "on" because I've used a wrong method
on the form data. Now it's equal to the value of the id of the selected
radio button.