Create loaded project's todos objects and DOM
The todos inside the project's arrays in localStorage are loaded, and
for each todo, an object is created for it and added to the todos array
of the corresponding project object instantiated from the Project class
(not the project object retrieved from localStorage).

All the todos DOMs are created as if defaultProject is pre-selected to
have its todos displayed (when in actuality, the currentProject variable
in `projectDOM.js` is yet to be defined, I tried assigning
defaultProject to it but that caused circular dependencies problems).