Solve todo edit issue
After applying the same solution as project edit, and the solution
from edit-problem branch, it didn't work.

The problem was in the naming of input fields (title was named title
and was retrieved using new_title) and in the retrieval of priority
input fields (should use #dialog-id followed by #priority).

Need to change ids to classes.