Fix todos not loading issue and duplicate todos
Todos not loading was due to the projects being re-added to localStorage
with an empty todos array each time the site reloaded.

After solving that problem, the default project would create 2 copies
of a new todo after reloading the page and submitting a todo add form.

The problem was due to defaultProject having 2 .addTodo methods in
the loading data section in index.js (projectObject.addTodo and
defaultProject.addTodo) while defaultProject was part of the iteration
of projectIDs.forEach